Move getLetterIconResId() from HomeScreen.kt to Utils.kt for future reuse
